Changed error update order of ops to match JLCS10.
This is for good reason, they put (rhat x db_j) at the
end to avoid accruing an extra round-off error for rhat.

Also increased `NUM_POINTS` from 400 to 401 in all remaining
places.

Note that the order of operations change has changed all plots
that use compensated de Casteljau (but not in a qualitative way).

I also went back and double-checked that `Table 1` was still
correct after the order of operations change.