Big update to `scripts/tests.cpp`.
The primary aim was to add tests for `de_casteljau::compensated_k` but
I also re-structured the tests file quite a bit along the way.

It **seems** there is a bug somewhere in the implementation (for K > 2),
but it could just be that small values are being printed as `0` or
`-0`. I need to follow up and print the binary representation
instead to figure out what's happening.